+++ Realm.Shared/linq/RealmResultsVisitor.cs
@@ -17,6 +17,7 @@
 ////////////////////////////////////////////////////////////////////////////
 
 using System;
+using System.Diagnostics;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
@@ -35,11 +36,12 @@ namespace Realms
         internal SortDescriptorBuilder OptionalSortDescriptorBuilder;  // set only when get OrderBy*
         private readonly RealmObject.Metadata _metadata;
 
-        private static class Methods 
+        private static class Methods
         {
             internal static LazyMethod Capture<T>(Expression<Action<T>> lambda)
             {
-                return new LazyMethod(() => {
+                return new LazyMethod(() =>
+                {
                     var method = (lambda.Body as MethodCallExpression).Method;
                     if (method.IsGenericMethod)
                     {
@@ -88,9 +90,10 @@ namespace Realms
         private void RecurseToWhereOrRunLambda(MethodCallExpression m)
         {
             this.Visit(m.Arguments[0]);  // creates the query or recurse to "Where"
-            if (m.Arguments.Count > 1) {
-                LambdaExpression lambda = (LambdaExpression)StripQuotes (m.Arguments[1]);
-                this.Visit (lambda.Body);
+            if (m.Arguments.Count > 1)
+            {
+                LambdaExpression lambda = (LambdaExpression)StripQuotes(m.Arguments[1]);
+                this.Visit(lambda.Body);
             }
         }
 
@@ -117,109 +120,163 @@ namespace Realms
         }
 
 
+        private RowHandle VisitElementAt(MethodCallExpression m)
+        {
+            Visit(m.Arguments.First());
+            var index = (int)ExtractConstantValue(m.Arguments.Last());
+
+            RowHandle row;
+            if (OptionalSortDescriptorBuilder == null)
+            {
+                var rowPtr = _coreQueryHandle.FindDirect((IntPtr)index);
+                row = Realm.CreateRowHandle(rowPtr, _realm.SharedRealmHandle);
+            }
+            else
+            {
+                using (ResultsHandle rh = _realm.MakeResultsForQuery(_coreQueryHandle, OptionalSortDescriptorBuilder))
+                {
+                    var rowPtr = rh.GetRow(index);
+                    row = Realm.CreateRowHandle(rowPtr, _realm.SharedRealmHandle);
+                }
+            }
+            return row;
+        }
+
+
         internal override Expression VisitMethodCall(MethodCallExpression m)
         {
             if (m.Method.DeclaringType == typeof(Queryable))
-            { 
-                if (m.Method.Name == "Where")
+            {
+                if (m.Method.Name == nameof(Queryable.Where))
                 {
                     this.Visit(m.Arguments[0]);
                     LambdaExpression lambda = (LambdaExpression)StripQuotes(m.Arguments[1]);
                     this.Visit(lambda.Body);
                     return m;
                 }
-                if (m.Method.Name == "OrderBy")
+                if (m.Method.Name == nameof(Queryable.OrderBy))
                 {
                     this.Visit(m.Arguments[0]);
                     AddSort((LambdaExpression)StripQuotes(m.Arguments[1]), true, true);
                     return m;
                 }
-                if (m.Method.Name == "OrderByDescending")
+                if (m.Method.Name == nameof(Queryable.OrderByDescending))
                 {
                     this.Visit(m.Arguments[0]);
                     AddSort((LambdaExpression)StripQuotes(m.Arguments[1]), true, false);
                     return m;
                 }
-                if (m.Method.Name == "ThenBy")
+                if (m.Method.Name == nameof(Queryable.ThenBy))
                 {
                     this.Visit(m.Arguments[0]);
                     AddSort((LambdaExpression)StripQuotes(m.Arguments[1]), false, true);
                     return m;
                 }
-                if (m.Method.Name == "ThenByDescending")
+                if (m.Method.Name == nameof(Queryable.ThenByDescending))
                 {
                     this.Visit(m.Arguments[0]);
                     AddSort((LambdaExpression)StripQuotes(m.Arguments[1]), false, false);
                     return m;
                 }
-                if (m.Method.Name == "Count")
+                if (m.Method.Name == nameof(Queryable.Count))
                 {
                     RecurseToWhereOrRunLambda(m);
                     var foundCount = _coreQueryHandle.Count();
                     return Expression.Constant(foundCount);
                 }
-                if (m.Method.Name == "Any")
+                if (m.Method.Name == nameof(Queryable.Any))
                 {
                     RecurseToWhereOrRunLambda(m);
                     bool foundAny = _coreQueryHandle.FindDirect(IntPtr.Zero) != IntPtr.Zero;
                     return Expression.Constant(foundAny);
                 }
-                if (m.Method.Name == "First")
+                if (m.Method.Name.StartsWith(nameof(Queryable.First)))
                 {
-                    RecurseToWhereOrRunLambda(m);  
+                    RecurseToWhereOrRunLambda(m);
                     IntPtr firstRowPtr = IntPtr.Zero;
                     if (OptionalSortDescriptorBuilder == null)
                     {
                         firstRowPtr = _coreQueryHandle.FindDirect(IntPtr.Zero);
                     }
-                    else 
+                    else
                     {
-                        using (ResultsHandle rh = _realm.MakeResultsForQuery(_coreQueryHandle, OptionalSortDescriptorBuilder)) 
+                        using (ResultsHandle rh = _realm.MakeResultsForQuery(_coreQueryHandle, OptionalSortDescriptorBuilder))
                         {
                             firstRowPtr = rh.GetRow(0);
                         }
                     }
-                    if (firstRowPtr == IntPtr.Zero)
+                    if (firstRowPtr != IntPtr.Zero)
+                        return Expression.Constant(_realm.MakeObjectForRow(_metadata, firstRowPtr));
+                    if (m.Method.Name == nameof(Queryable.First))
                         throw new InvalidOperationException("Sequence contains no matching element");
-                    return Expression.Constant(_realm.MakeObjectForRow(_metadata, firstRowPtr));
+                    Debug.Assert(m.Method.Name == nameof(Queryable.FirstOrDefault));
+                    return Expression.Constant(null);
                 }
-                if (m.Method.Name == "Single")  // same as unsorted First with extra checks
+/*
+// FIXME: See discussion in the test DefaultIfEmptyReturnsDefault
+// kept because it shows part of what might be a viable implementation if can work out architectural issues
+
+                if (m.Method.Name == nameof(Queryable.DefaultIfEmpty))
                 {
-                    RecurseToWhereOrRunLambda(m);  
+                    RecurseToWhereOrRunLambda(m);
+                    IntPtr firstRowPtr = _coreQueryHandle.FindDirect(IntPtr.Zero);
+                    if (firstRowPtr != IntPtr.Zero)
+                        return m;  // as if just a "Where"
+                    var innerType = m.Type.GetGenericArguments()[0];
+                    var listType = typeof(List<>).MakeGenericType(innerType);
+                    var singleNullItemList = Activator.CreateInstance(listType);
+                    ((IList)singleNullItemList).Add(null);
+                    return Expression.Constant(singleNullItemList);
+                }
+*/
+                if (m.Method.Name.StartsWith(nameof(Queryable.Single)))  // same as unsorted First with extra checks
+                {
+                    RecurseToWhereOrRunLambda(m);
                     var firstRowPtr = _coreQueryHandle.FindDirect(IntPtr.Zero);
                     if (firstRowPtr == IntPtr.Zero)
-                        throw new InvalidOperationException("Sequence contains no matching element");
+                    {
+                        if (m.Method.Name == nameof(Queryable.Single))
+                            throw new InvalidOperationException("Sequence contains no matching element");
+                        Debug.Assert(m.Method.Name == nameof(Queryable.SingleOrDefault));
+                        return Expression.Constant(null);
+                    }
                     var firstRow = Realm.CreateRowHandle(firstRowPtr, _realm.SharedRealmHandle);
-                    IntPtr nextIndex = (IntPtr)(firstRow.RowIndex+1);
+                    IntPtr nextIndex = (IntPtr)(firstRow.RowIndex + 1);
                     var nextRowPtr = _coreQueryHandle.FindDirect(nextIndex);
                     if (nextRowPtr != IntPtr.Zero)
                         throw new InvalidOperationException("Sequence contains more than one matching element");
                     return Expression.Constant(_realm.MakeObjectForRow(_metadata, firstRow));
                 }
-                if (m.Method.Name == nameof(Queryable.ElementAt))
+                if (m.Method.Name.StartsWith(nameof(Queryable.Last)))
                 {
-                    Visit(m.Arguments.First());
-                    var index = (int)ExtractConstantValue(m.Arguments.Last());
+                    RecurseToWhereOrRunLambda(m);
 
-                    RowHandle row = null;
-                    if (OptionalSortDescriptorBuilder == null)
+                    var lastRowPtr = IntPtr.Zero;
+                    using (ResultsHandle rh = _realm.MakeResultsForQuery(_coreQueryHandle, OptionalSortDescriptorBuilder))
                     {
-                        var rowPtr = _coreQueryHandle.FindDirect((IntPtr)index);
-                        row = Realm.CreateRowHandle(rowPtr, _realm.SharedRealmHandle);
+                        var lastIndex = rh.Count() - 1;
+                        if (lastIndex >= 0)
+                            lastRowPtr = rh.GetRow(lastIndex);
                     }
-                    else
-                    {
-                        using (ResultsHandle rh = _realm.MakeResultsForQuery(_coreQueryHandle, OptionalSortDescriptorBuilder))
-                        {
-                            var rowPtr = rh.GetRow(index);
-                            row = Realm.CreateRowHandle(rowPtr, _realm.SharedRealmHandle);
-                        }
-                    }
+                    if (lastRowPtr != IntPtr.Zero)
+                        return Expression.Constant(_realm.MakeObjectForRow(_metadata, lastRowPtr));
+                    if (m.Method.Name == nameof(Queryable.Last))
+                        throw new InvalidOperationException("Sequence contains no matching element");
+                    Debug.Assert(m.Method.Name == nameof(Queryable.LastOrDefault));
+                    return Expression.Constant(null);
+                }
+                if (m.Method.Name.StartsWith(nameof(Queryable.ElementAt)))
+                {
+                    var row = VisitElementAt(m);
                     if (row == null || row.IsInvalid)
-                        throw new IndexOutOfRangeException();
+                    {
+                        if (m.Method.Name == nameof(Queryable.ElementAt))
+                            throw new ArgumentOutOfRangeException();
+                        Debug.Assert(m.Method.Name == nameof(Queryable.ElementAtOrDefault));
+                        return Expression.Constant(null);
+                    }
                     return Expression.Constant(_realm.MakeObjectForRow(_metadata, row));
                 }
-
             }
 
             if (m.Method.DeclaringType == typeof(string))
@@ -248,7 +305,7 @@ namespace Realms
                     }
                     var columnIndex = _coreQueryHandle.GetColumnIndex(member.Member.Name);
 
-                    var argument = ExtractConstantValue (m.Arguments.SingleOrDefault());
+                    var argument = ExtractConstantValue(m.Arguments.SingleOrDefault());
                     if (argument == null || argument.GetType() != typeof(string))
                     {
                         throw new NotSupportedException($"The method '{m.Method}' has to be invoked with a single string constant argument or closure variable");
@@ -265,11 +322,11 @@ namespace Realms
         {
             switch (u.NodeType)
             {
-            case ExpressionType.Not:
-                {
+                case ExpressionType.Not:
+                    {
                         _coreQueryHandle.Not();
-                    this.Visit (u.Operand);  // recurse into richer expression, expect to VisitCombination
-                }
+                        this.Visit(u.Operand);  // recurse into richer expression, expect to VisitCombination
+                    }
                     break;
                 default:
                     throw new NotSupportedException($"The unary operator '{u.NodeType}' is not supported");
@@ -277,7 +334,7 @@ namespace Realms
             return u;
         }
 
-        protected void VisitCombination(BinaryExpression b,  Action<QueryHandle> combineWith )
+        protected void VisitCombination(BinaryExpression b, Action<QueryHandle> combineWith)
         {
             _coreQueryHandle.GroupBegin();
             Visit(b.Left);
@@ -300,7 +357,7 @@ namespace Realms
                 // handle closure variables
                 return ((System.Reflection.FieldInfo)memberAccess.Member).GetValue(((ConstantExpression)memberAccess.Expression).Value);
             }
-                
+
             return null;
         }
 
@@ -308,7 +365,7 @@ namespace Realms
         {
             if (b.NodeType == ExpressionType.AndAlso)  // Boolean And with short-circuit
             {
-                VisitCombination(b, (qh) => { /* noop -- AND is the default combinator */} );
+                VisitCombination(b, (qh) => { /* noop -- AND is the default combinator */});
             }
             else if (b.NodeType == ExpressionType.OrElse)  // Boolean Or with short-circuit
             {
@@ -317,10 +374,24 @@ namespace Realms
             else
             {
                 var leftMember = b.Left as MemberExpression;
+                string leftName = null;  // yes you need to init this for the if statement below
                 if (leftMember == null)
-                    throw new NotSupportedException(
-                        $"The lhs of the binary operator '{b.NodeType}' should be a member expression. \nUnable to process `{b.Left}`");
-                var leftName = leftMember.Member.Name;
+                {
+                    // bit of a hack to cope with the way LINQ changes the RHS of a char literal to an Int32
+                    // so an incoming lambda looks like {p => (Convert(p.CharProperty) == 65)}
+                    // from Where(p => p.CharProperty == 'A')
+                    var leftConvert = b.Left as UnaryExpression;
+                    if (leftConvert?.NodeType == ExpressionType.Convert)
+                    {
+                        var leftConvertMember = leftConvert.Operand as MemberExpression;
+                        leftName = leftConvertMember?.Member.Name;
+                    }
+                    if (leftName == null)
+                        throw new NotSupportedException(
+                            $"The lhs of the binary operator '{b.NodeType}' should be a member expression. \nUnable to process `{b.Left}`");
+                }
+                else
+                    leftName = leftMember.Member.Name;
 
                 var rightValue = ExtractConstantValue(b.Right);
                 if (rightValue == null)
@@ -362,13 +433,15 @@ namespace Realms
         }
 
         private static void AddQueryEqual(QueryHandle queryHandle, string columnName, object value)
-            {
+        {
             var columnIndex = queryHandle.GetColumnIndex(columnName);
 
             if (value is string)
                 queryHandle.StringEqual(columnIndex, (string)value);
             else if (value is bool)
                 queryHandle.BoolEqual(columnIndex, (bool)value);
+            else if (value is char)
+                queryHandle.IntEqual(columnIndex, (int)value);
             else if (value is int)
                 queryHandle.IntEqual(columnIndex, (int)value);
             else if (value is long)
@@ -409,6 +482,8 @@ namespace Realms
                 queryHandle.StringNotEqual(columnIndex, (string)value);
             else if (value is bool)
                 queryHandle.BoolNotEqual(columnIndex, (bool)value);
+            else if (value is char)
+                queryHandle.IntNotEqual(columnIndex, (int)value);
             else if (value is int)
                 queryHandle.IntNotEqual(columnIndex, (int)value);
             else if (value is long)
@@ -419,7 +494,7 @@ namespace Realms
                 queryHandle.DoubleNotEqual(columnIndex, (double)value);
             else if (value is DateTimeOffset)
                 queryHandle.TimestampTicksNotEqual(columnIndex, (DateTimeOffset)value);
-            else if (value.GetType()== typeof(byte[]))
+            else if (value.GetType() == typeof(byte[]))
             {
                 var buffer = (byte[])value;
                 if (buffer.Length == 0)
@@ -445,7 +520,9 @@ namespace Realms
         {
             var columnIndex = queryHandle.GetColumnIndex(columnName);
 
-            if (value is int)
+            if (value is char)
+                queryHandle.IntLess(columnIndex, (int)value);
+            else if (value is int)
                 queryHandle.IntLess(columnIndex, (int)value);
             else if (value is long)
                 queryHandle.LongLess(columnIndex, (long)value);
@@ -465,7 +542,9 @@ namespace Realms
         {
             var columnIndex = queryHandle.GetColumnIndex(columnName);
 
-            if (value is int)
+            if (value is char)
+                queryHandle.IntLessEqual(columnIndex, (int)value);
+            else if (value is int)
                 queryHandle.IntLessEqual(columnIndex, (int)value);
             else if (value is long)
                 queryHandle.LongLessEqual(columnIndex, (long)value);
@@ -485,7 +564,9 @@ namespace Realms
         {
             var columnIndex = queryHandle.GetColumnIndex(columnName);
 
-            if (value is int)
+            if (value is char)
+                queryHandle.IntGreater(columnIndex, (int)value);
+            else if (value is int)
                 queryHandle.IntGreater(columnIndex, (int)value);
             else if (value is long)
                 queryHandle.LongGreater(columnIndex, (long)value);
@@ -505,7 +586,9 @@ namespace Realms
         {
             var columnIndex = queryHandle.GetColumnIndex(columnName);
 
-            if (value is int)
+            if (value is char)
+                queryHandle.IntGreaterEqual(columnIndex, (int)value);
+            else if (value is int)
                 queryHandle.IntGreaterEqual(columnIndex, (int)value);
             else if (value is long)
                 queryHandle.LongGreaterEqual(columnIndex, (long)value);
@@ -540,11 +623,11 @@ namespace Realms
             {
                 if (c.Value is bool)
                 {
-                } 
+                }
                 else if (c.Value is string)
                 {
                 }
-                else if (c.Value.GetType() == typeof (object))
+                else if (c.Value.GetType() == typeof(object))
                 {
                     throw new NotSupportedException($"The constant for '{c.Value}' is not supported");
                 }
@@ -577,7 +660,8 @@ namespace Realms
         {
             if (m.Expression != null && m.Expression.NodeType == ExpressionType.Parameter)
             {
-                if (m.Type == typeof(bool)) {
+                if (m.Type == typeof(bool))
+                {
                     object rhs = true;  // box value
                     var leftName = m.Member.Name;
                     AddQueryEqual(_coreQueryHandle, leftName, rhs);